+++ src/main/java/lotto/model/LottoNumberRepositoryImpl.java
@@ -2,11 +2,11 @@ package lotto.model;
 
 import java.util.ArrayList;
 import java.util.List;
-import lotto.Lotto;
+import lotto.model.entity.Lotto;
 
 public class LottoNumberRepositoryImpl implements LottoNumberRepository {
 
-    private final List<Lotto> lottoList = new ArrayList<Lotto>();
+    private final List<Lotto> lottoList = new ArrayList<>();
 
     @Override
     public void saveLotto(Lotto lotto) {
